event handlers [nfc]: Move orientation tracking into a handler-component.
