+++ src/boot/AppEventHandlers.js
@@ -3,7 +3,6 @@
 import React, { PureComponent } from 'react';
 import { AppState, View, StyleSheet, Platform, NativeModules } from 'react-native';
 import SafeArea from 'react-native-safe-area';
-import Orientation, { type Orientations } from 'react-native-orientation';
 
 import type { Node as React$Node } from 'react';
 import type { Dispatch } from '../types';
@@ -14,9 +13,10 @@ import {
   NotificationListener,
   notificationOnAppActive,
 } from '../notification';
-import { appOrientation, initSafeAreaInsets } from '../actions';
+import { initSafeAreaInsets } from '../actions';
 import PresenceHeartbeat from '../presence/PresenceHeartbeat';
 import NetworkStateHandler from './event-handlers/NetworkStateHandler';
+import OrientationStateHandler from './event-handlers/OrientationHandler';
 
 const styles = StyleSheet.create({
   wrapper: {
@@ -33,26 +33,6 @@ type Props = $ReadOnly<{|
 |}>;
 
 class AppEventHandlers extends PureComponent<Props> {
-  // The libdef says that this parameter is optional. The docs disagree.
-  handleOrientationChange = (orientation?: Orientations) => {
-    if (!orientation) {
-      return;
-    }
-
-    // Lookup table. More elaborate than a simple 'orientation !== LANDSCAPE'
-    // test to ensure robustness against future additions to the enumeration.
-    const converter = {
-      LANDSCAPE: 'LANDSCAPE',
-      PORTRAIT: 'PORTRAIT',
-      PORTRAITUPSIDEDOWN: 'PORTRAIT',
-      UNKNOWN: 'PORTRAIT',
-    };
-    const converted = converter[orientation];
-
-    const { dispatch } = this.props;
-    dispatch(appOrientation(converted));
-  };
-
   /** For the type, see docs: https://facebook.github.io/react-native/docs/appstate */
   handleAppStateChange = (state: 'active' | 'background' | 'inactive') => {
     const { unreadCount } = this.props;
@@ -74,13 +54,11 @@ class AppEventHandlers extends PureComponent<Props> {
     SafeArea.getSafeAreaInsetsForRootView().then(params =>
       dispatch(initSafeAreaInsets(params.safeAreaInsets)),
     );
-    Orientation.addOrientationListener(this.handleOrientationChange);
     this.notificationListener.start();
   }
 
   componentWillUnmount() {
     AppState.removeEventListener('change', this.handleAppStateChange);
-    Orientation.removeOrientationListener(this.handleOrientationChange);
     this.notificationListener.stop();
   }
 
@@ -88,6 +66,7 @@ class AppEventHandlers extends PureComponent<Props> {
     return (
       <>
         <NetworkStateHandler />
+        <OrientationStateHandler />
         <PresenceHeartbeat />
         <View style={styles.wrapper}>{this.props.children}</View>
       </>
